requirements.txt
annotated-types==0.7.0
anyio==4.4.0
appnope==0.1.4
asttokens==2.4.1
attrs==23.2.0
beautifulsoup4==4.12.3
blis==0.7.11
Brotli==1.0.9
bs4==0.0.2
catalogue==2.0.10
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.0.3
cloudpathlib==0.19.0
colorama==0.4.4
comm==0.2.2
confection==0.1.5
contourpy==1.2.1
cycler==0.12.1
cymem==2.0.8
debugpy==1.8.1
decorator==5.1.1
fastapi==0.115.0
fastjsonschema==2.20.0
Flask==2.0.2
Flask-Compress==1.10.1
fonttools==4.53.1
gensim==4.3.3
gunicorn==20.1.0
h11==0.14.0
idna==3.8
ipykernel==6.29.4
ipython==8.24.0
itsdangerous==2.0.1
jedi==0.19.1
Jinja2==3.0.3
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
langcodes==3.4.0
language_data==1.2.0
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.0.1
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
murmurhash==1.0.10
nbformat==5.10.4
nest-asyncio==1.6.0
nltk==3.6.6
numpy==1.26.4
packaging==24.0
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.2
plotly==5.5.0
preshed==3.0.9
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.9.1
pydantic_core==2.23.3
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.8.2
pytz==2021.3
pyzmq==26.0.3
rapidfuzz==3.9.7
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
rich==13.8.1
rpds-py==0.19.0
scikit-learn==1.5.0
scipy==1.13.1
seaborn==0.13.2
setuptools==70.0.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
soupsieve==2.6
spacy==3.7.6
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.35
srsly==2.4.8
stack-data==0.6.3
starlette==0.38.5
tenacity==8.0.1
thefuzz==0.22.1
thinc==8.2.5
threadpoolctl==3.5.0
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
uvicorn==0.30.6
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
Werkzeug==2.0.2
wheel==0.43.0
wordcloud==1.9.3
wrapt==1.16.0
xgboost==2.1.0
